+++ Point-of-No-Return/UI/ActionUI/ActionUI.h
@@ -1,159 +1,13 @@
 ﻿#ifndef ACTIONUI_H
 #define ACTIONUI_H
 
-#include"../UI.h"
+#include "../UI.h"
+#include "../ActionUI/BarBase.h"
+#include "../ActionUI/FrameBase.h"
+#include "../ActionUI/IconBase.h"
 
 namespace ActionUI
 {
-	/**
-	 * @brief BerBaseクラス
-	 */
-	class BarBase
-	{
-	public:
-
-		DX::DirectX& dx = DX::DirectX::GetInstance();
-
-		virtual void Initialize() {};
-		virtual void UpDate() {};
-		virtual void Draw() {};
-		BarBase() {};
-		virtual ~BarBase() {};
-
-
-	protected:
-
-		Vec2 position =
-		{
-			CoordinateX(0),
-			CoordinateY(0)
-		};
-
-		Size size =
-		{
-			Width(0),
-			Height(0)
-		};
-
-		TextureUV uv =
-		{
-			TextureU(0),
-			TextureV(0)
-		};
-
-		TextureSize texture_size =
-		{
-			Width(0),
-			Height(0)
-		};
-
-		TextureName texture_name = TextureName("Action");
-
-		Degree degree = Degree(0);
-		Zoom zoom = Zoom(0);
-		IsReverse is_reverse = IsReverse(false);
-
-	
-	};
-
-	/**
-	 * @brief FrameBaseクラス
-	 */
-	class FrameBase
-	{
-	public:
-
-		DX::DirectX& dx = DX::DirectX::GetInstance();
-
-		virtual void Initialize() {};
-		virtual void Draw() {};
-		FrameBase() {};
-		virtual ~FrameBase() {};
-
-	protected:
-
-		Vec2 position =
-		{
-			CoordinateX(0),
-			CoordinateY(0)
-		};
-
-		Size size =
-		{
-			Width(0),
-			Height(0)
-		};
-
-		TextureUV uv =
-		{
-			TextureU(0),
-			TextureV(0)
-		};
-
-		TextureSize texture_size =
-		{
-			Width(0),
-			Height(0)
-		};
-
-		TextureName texture_name = TextureName("Action");
-
-		Degree degree = Degree(0);
-		Zoom zoom = Zoom(0);
-		IsReverse is_reverse = IsReverse(false);
-
-	};
-
-	/**
-	 * @brief ItemIconBaseクラス
-	 */
-	class IconBase
-	{
-	public:
-
-		DX::DirectX& dx = DX::DirectX::GetInstance();
-
-		virtual void Initialize() {};
-		virtual void Draw() {};
-
-		IconBase() {};
-		virtual ~IconBase() {};
-
-	protected:
-
-		Vec2 position =
-		{
-			CoordinateX(0),
-			CoordinateY(0)
-		};
-
-		Size size =
-		{
-			Width(0),
-			Height(0)
-		};
-
-		TextureUV uv =
-		{
-			TextureU(0),
-			TextureV(0)
-		};
-
-		TextureSize texture_size =
-		{
-			Width(0),
-			Height(0)
-		};
-
-		TextureName texture_name = TextureName("Action");
-
-		Degree degree = Degree(0);
-		Zoom zoom = Zoom(0);
-		IsReverse is_reverse = IsReverse(false);
-
-
-	};
-
 	/**
 	 * @brief ItemUIクラス
 	 */
